+++ src/main/java/com/hmdandelion/project_1410002/sales/domain/type/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.hmdandelion.project_1410002.sales.model;
+package com.hmdandelion.project_1410002.sales.domain.type;
 
 public enum OrderStatus {
     COMPLETED, ORDER_RECEIVED
